Library compatibility spec for bibtex-ruby
In order to ensure that future versions of the
bibtex-ruby gem do not introduce API changes that
break compatibility, this commit adds a spec that
compares a hash constructed by bibtex-ruby from
a .bib file to the hash I expect.

This is also practical for visualizing how bibtex-ruby
converts BibTeX values to citeproc format, especially
for names and dates.