Improves Readme and Style::FIELDS documentation
The Readme file has been updated a lot and now contains
a (more or less) complete style scripting guide as well
as general usage instructions etc. It is now added to
the documentation automatically per the new option in
.yardopts. For standards compliance, the file has been
renamed to 'README'.

Moreover, documentation of Style::FIELDS has been greatly
improved and the teststyle has been updated to reflect
API changes introduced with the last commits.