Gemification
1) Directory Tree

RCite now uses the standard gem directory structure with
lib files going into lib/rcite/ and the loader file
being lib/rcite.rb.

2) New Load Structure

Due to the updated directory tree, 'require' statements
had to be changed in a whole bunch of files.

3) rcite.gemspec

The gemspec file now provides all relevant information about
the gem. Note that the dependency management has been moved
from Gemfile to rcite.gemspec, with Gemfile referencing the
latter. Using bundler is of course still possible.

Gemfile.lock is now excluded from the git repo because I
realised that it doesn't make a lot of sense for a gem.
Head to the following article for an explanation:
http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/