More robust style loading
In certain test scenarios, the previous method of loading styles
using `require` could be problematic. It wouldn't load the same
file twice even if that file had been modified by a testing
routine but rather continue without indicating an error.

Therefore, `require` has been replaced by `load` together with
a statement that throws an error if a style has already been
loaded.

The Teststyle spec had to be modified slightly because of this
change.

closes #12