Adds features for multiciting and the bibliography
Multiciting (which can actually be multibib'ing as well)
stands for the concept of citing/bib'ing multiple texts
with one command in preprocessing mode.

In addition to that, a new command called 'bibliography'
will be added to the preprocessing mode. This command
will automatically create a bibliography from all the
texts that were cited in the file.