Preparations for new option interface
The new Option class will be the main component of a new option
management system. Currently, there is an unfortunate tension
between the `defaults` hash that controls the generation of
the author/editor list and the new `around`/`between` methods
that reside in the style preamble. These two 'solutions' will hence be
replaced by a unified option management.